Move the booksSeed function to \seed\booksSeed.js
